Refactor event detail to split UI and logic
