+++ src/screens/eventDetail/EventDetailsViewModel.tsx
@@ -1,4 +1,3 @@
-import * as AddCalendarEvent from 'react-native-add-calendar-event'
 import { PollRowViewModel } from '../polls/PollRowViewModel'
 
 export interface EventDetailsViewModel {
@@ -14,7 +13,7 @@ export interface EventDetailsViewModel {
   date: EventDateViewModel
   eventUrl?: string
   description: string
-  calendarEvent: AddCalendarEvent.CreateOptions
+  canSeeMore: boolean
   survey?: PollRowViewModel
 }
 
@@ -27,9 +26,4 @@ export interface EventOrganizerViewModel {
   title: string
   description?: string
   openUrl?: string
-}
-
-export interface EventDescriptionViewModel {
-  description: string
-  canSeeMore: boolean
 }
